+++ src/components/button/icon-button.tsx
@@ -1,5 +1,5 @@
 import { LucideIcon } from "lucide-react";
-import { Button, ButtonProps } from "./ui/button";
+import { Button, ButtonProps } from "../ui/button";
 
 interface IconButtonProps extends ButtonProps {
   Icon: LucideIcon;
